Id type changed to Integer.

+++ src/main/java/com/cocaine/practice_app/blackjack_java_reboot/games/repository/GameEntity.java
@@ -13,6 +13,6 @@ public final class GameEntity {
 
     @Id
     @Getter
-    private final Long Id;
+    private final Integer Id;
 
 }
